Update sampler to allow float factors, ie sampling 0.1% of data
git-svn-id: http://www.mpa-garching.mpg.de/svn/cosmo-group/kdolag/splotch@26592 019ef673-9300-0410-806d-f16013bd1951

+++ reader/gadget_reader.cc
@@ -59,7 +59,7 @@ int gadget_find_block (bifstream &file,const string &label)
   return(blocksize-8);
 }
 
-int gadget_read_header(bifstream &file, unsigned int *npart, double &time, unsigned int *nparttotal, double &boxsize, paramfile &params, int sample_factor)
+int gadget_read_header(bifstream &file, unsigned int *npart, double &time, unsigned int *nparttotal, double &boxsize, paramfile &params, float sample_factor)
 {
   bool doSample = params.find<bool>("sampler",false);
   double h,O,L;
@@ -106,11 +106,11 @@ void gadget_reader(paramfile &params, int interpol_mode,
   // Sampling, this cannot be done while interpolating...
   // Sample factor is read in as percentage then converted to factor
   bool doSample = params.find<bool>("sampler",false);
-  int sample_factor = 1;
+  float sample_factor = 1;
   if(doSample && !interpol_mode)
   { 
     // If no factor, use 100% sampling ie no sample.
-    sample_factor = params.find<int>("sample_factor",100);
+    sample_factor = params.find<float>("sample_factor",100);
     if((sample_factor > 0) && (sample_factor <= 100))
     {
       sample_factor = 100/sample_factor;
